Move squad strength calculation from Finance to Squad module
calculateLeagueStrengths, calculateSquadStrength, and getProjectedPosition
now live in SquadService where they belong. The strength formula is
simplified to use only technical + physical ability, removing the
dependency on volatile match state (fitness/morale). Also removes the
dead calculateSquadStrength method from BudgetProjectionService.
